test: Move non-port cell-include cursor tests next to the code
Keep parser/src/tests/asciidoctor_rb/tables_test.rs as a strict mirror of the
ported Ruby Asciidoctor tests: the faithful port of `tables_test.rb` 1728 stays
there, while the two additional regression tests for the same fix (the
root-document case and the nested-table guard) move into the inline
`#[cfg(test)] mod tests` in blocks/table.rs, next to the code under test.

No behavior change.
